requirements.txt
datasets>=1.11.0
flatten-dict
hydra-core>=1.0.6
isort
librosa
pycocotools
pydicom
pylint
pytorch-lightning
scipy~=1.13.0
toml
torch
torchaudio
torchmetrics
torchvision
tqdm
transformers
wandb
wget
yapf
einops~=0.8.0
hub
wfdb
timm
h5py
pylidc
bs4
mne
SimpleITK
ecg-plot
numpy
opencv-python
pandas
scikit-learn
beautifulsoup4
matplotlib
pandas
pillow
nibabel